feat(nonenull): improve type signature for noneNull
It is now possible to pass a tuple type to noneNull and get back the appropriate tuple type.

BREAKING CHANGE: The type parameter to noneNull has changed to the type of the array itself, not the
type of its elements. The behaviour of the function is unchanged. Callers that specify an explicit
type parameter will need to be updated.